requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
spotipy==2.22.1
yt-dlp>=2024.1.0
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0 
